Use current working directory in read cmd (#6678)
That is, instead of defaulting to the process' working directory (VS
Code), use the opened workspace or file for the working directory as Vim
itself does.
- If a workspace folder is opened in Code, use that
- If no workspace folder is open, use the active editor's file's folder
- If neither is available, default to the process' working directory

One issue still remaining here: if the user has multiple workspaces
running, only the first one will be used. This edge case should be
handled.